+++ brain.go
@@ -7,6 +7,7 @@ import (
 	"runtime"
 	"strings"
 	"sync"
+	"sync/atomic"
 	"time"
 
 	"github.com/pkg/errors"
@@ -24,11 +25,16 @@ type Brain struct {
 	mu     sync.RWMutex // mu protects concurrent access to the Memory
 	memory Memory
 
-	events         chan Event
+	eventsInput chan Event // input for any new events, the Brain ensures that callers never block when writing to it
+	eventsLoop  chan Event // used in Brain.HandleEvents() to actually process the events
+	shutdown    chan shutdownRequest
+
 	handlers       map[reflect.Type][]eventHandler
-	handlerTimeout time.Duration // zero means no timeout
+	handlerTimeout time.Duration // zero means no timeout, defaults to one minute
 
 	registrationErrs []error // any errors that occurred during setup (e.g. in Bot.RegisterHandler)
+	handlingEvents   int32   // accessed atomically (non-zero means the event handler was started)
+	closed           int32   // accessed atomically (non-zero means the brain was shutdown already)
 }
 
 // An Event represents a concrete event type and optional callbacks that are
@@ -38,46 +44,48 @@ type Event struct {
 	Callbacks []func(Event)
 }
 
-// An event handler is a function that takes a context and the reflected value
+// The shutdownRequest type is used when signaling shutdown information between
+// Brain.Shutdown() and the Brain.HandleEvents loop.
+type shutdownRequest struct {
+	ctx      context.Context
+	callback chan bool
+}
+
+// An eventHandler is a function that takes a context and the reflected value
 // of a concrete event type.
 type eventHandler func(context.Context, reflect.Value) error
-
-// The EventRegistry is the interface that is exposed to Adapter implementations
-// when connecting to the Brain. Note that this interface actually exposes direct
-// write access to the events channel to allow adapters to deliver events
-// synchronously and in deterministic order.
-type EventRegistry interface {
-	Channel() chan<- Event
-	RegisterHandler(function interface{})
-}
-
-// brainRegistry implements the EventRegistry to connect a Brain with its Adapter.
-type brainRegistry struct {
-	*Brain
-}
-
-// Channel returns the events channel of the brain.
-func (a brainRegistry) Channel() chan<- Event {
-	return a.events
-}
 
 // NewBrain creates a new robot Brain. By default the Brain will use a Memory
 // implementation that stores all keys and values directly in memory. You can
 // change the memory implementation afterwards by simply assigning to
 // Brain.Memory. If the passed logger is nil it will fallback to the
-// zap.NewNop() logger. By default no timeout will be enforced on the event
-// handlers.
+// zap.NewNop() logger.
 func NewBrain(logger *zap.Logger) *Brain {
 	if logger == nil {
 		logger = zap.NewNop()
 	}
 
-	return &Brain{
-		logger:   logger,
-		memory:   newInMemory(),
-		events:   make(chan Event, 10),
-		handlers: make(map[reflect.Type][]eventHandler),
+	b := &Brain{
+		logger:         logger,
+		memory:         newInMemory(),
+		eventsInput:    make(chan Event),
+		eventsLoop:     make(chan Event),
+		shutdown:       make(chan shutdownRequest),
+		handlers:       make(map[reflect.Type][]eventHandler),
+		handlerTimeout: time.Minute,
 	}
+
+	b.consumeEvents()
+
+	return b
+}
+
+func (b *Brain) isHandlingEvents() bool {
+	return atomic.LoadInt32(&b.handlingEvents) == 1
+}
+
+func (b *Brain) isClosed() bool {
+	return atomic.LoadInt32(&b.closed) == 1
 }
 
 // RegisterHandler registers a function to be executed when a specific event is
@@ -141,40 +149,121 @@ func (b *Brain) registerHandler(fun interface{}) error {
 	return nil
 }
 
-func (b *Brain) connectAdapter(a Adapter) {
-	a.Register(brainRegistry{b})
-}
-
 // Emit sends the first argument as event to the brain from where it is
-// dispatched to all registered handlers.
+// dispatched to all registered handlers. This function blocks until the event
+// handler was started via Brain.HandleEvents(). When the event handler is
+// running it will return immediately.
 func (b *Brain) Emit(event interface{}, callbacks ...func(Event)) {
-	go func() {
-		b.events <- Event{Data: event, Callbacks: callbacks}
-	}()
+	if b.isClosed() {
+		b.logger.Debug(
+			"Ignoring new event because brain is currently shutting down or is already closed",
+			zap.String("type", fmt.Sprintf("%T", event)),
+		)
+		return
+	}
+
+	b.eventsInput <- Event{Data: event, Callbacks: callbacks}
 }
 
 // HandleEvents starts the event handler loop of the Brain. This function blocks
-// until the passed context is canceled. If no handler timeout was configured
-// the brain might block indefinitely even if the context is canceled but an
-// event handler or callback is not respecting the context.
-func (b *Brain) HandleEvents(ctx context.Context) {
+// until Brain.Shutdown() is called and returned.
+func (b *Brain) HandleEvents() {
+	if b.isClosed() {
+		b.logger.Error("HandleEvents failed because bot is already closed")
+		return
+	}
+
+	ctx := context.Background()
+	var shutdown shutdownRequest // set when Brain.Shutdown() is called
+
+	atomic.StoreInt32(&b.handlingEvents, 1)
 	b.handleEvent(ctx, Event{Data: InitEvent{}})
 
 	for {
 		select {
-		case evt := <-b.events:
+		case evt, ok := <-b.eventsLoop:
+			if !ok {
+				// Brain.consumeEvents() is done processing all remaining events
+				// and we can now safely shutdown the event handler, knowing that
+				// all pending events have been processed.
+				b.handleEvent(ctx, Event{Data: ShutdownEvent{}})
+				shutdown.callback <- true
+				return
+			}
+
 			b.handleEvent(ctx, evt)
 
-		case <-ctx.Done():
-			b.handleEvent(ctx, Event{Data: ShutdownEvent{}})
-			return
+		case shutdown = <-b.shutdown:
+			// The Brain is shutting down. We have to close the input channel so
+			// we doe no longer accept new events and only process the remaining
+			// pending events. When the goroutine of Brain.consumeEvents() is
+			// done it will close the events loop channel and the case above will
+			// use the shutdown callback and return from this function.
+			ctx = shutdown.ctx
+			close(b.eventsInput)
+			atomic.StoreInt32(&b.handlingEvents, 0)
 		}
 	}
 }
 
-// handleEvent receives an event and determines which handler it must be
-// dispatched to using the reflect API. Additionally the function enforces any
-// event handler timeouts (if configured) and runs any event callbacks.
+// consumeEvents continuously reads events from b.eventsInput in a new goroutine
+// so emitting an event never blocks on the caller. All events will be returned
+// in the result channel of this function in the same order in which they have
+// been inserted into b.events. In this sense this function provides an events
+// channel with "infinite" capacity. The spawned goroutine stops when the
+// b.eventsInput channel is closed.
+func (b *Brain) consumeEvents() {
+	var queue []Event
+	b.eventsLoop = make(chan Event)
+
+	outChan := func() chan Event {
+		if len(queue) == 0 {
+			// In case the queue is empty we return a nil channel to disable the
+			// corresponding select case in the goroutine below.
+			return nil
+		}
+
+		return b.eventsLoop
+	}
+
+	nextEvt := func() Event {
+		if len(queue) == 0 {
+			// Prevent index out of bounds if there is no next event. Note that
+			// this event is actually never received because the outChan()
+			// function above will return "nil" in this case which disables the
+			// corresponding select case.
+			return Event{}
+		}
+
+		return queue[0]
+	}
+
+	go func() {
+		for {
+			select {
+			case evt, ok := <-b.eventsInput:
+				if !ok {
+					// Events input channel was closed because Brain is shutting
+					// down. Emit all pending events from the queue and then close
+					// the events loop channel so Brain.HandleEvents() can exit.
+					for _, evt := range queue {
+						b.eventsLoop <- evt
+					}
+					close(b.eventsLoop)
+					return
+				}
+
+				queue = append(queue, evt)
+			case outChan() <- nextEvt(): // disabled if len(queue) == 0
+				queue = queue[1:]
+			}
+		}
+	}()
+}
+
+// handleEvent receives an event and dispatches it to all registered handlers
+// using the reflect API. When all applicable handlers are called (maybe none)
+// the function runs all event callbacks.
 func (b *Brain) handleEvent(ctx context.Context, evt Event) {
 	event := reflect.ValueOf(evt.Data)
 	typ := event.Type()
@@ -193,8 +282,6 @@ func (b *Brain) handleEvent(ctx context.Context, evt Event) {
 		}
 	}
 
-	// TODO: callbacks should also get a context
-	// TODO: respect context even if callbacks don't
 	for _, callback := range evt.Callbacks {
 		callback(evt)
 	}
@@ -218,6 +305,50 @@ func (b *Brain) executeEventHandler(ctx context.Context, handler eventHandler, e
 	case <-ctx.Done():
 		return ctx.Err()
 	}
+}
+
+// Shutdown stops event handler loop of the Brain and waits until all pending
+// events have been processed. After the brain is shutdown, it will no longer
+// accept new events. The passed context can be used to stop waiting for any
+// pending events or handlers and instead exit immediately (e.g. after a timeout
+// or a second SIGTERM).
+func (b *Brain) Shutdown(ctx context.Context) {
+	closing := atomic.CompareAndSwapInt32(&b.closed, 0, 1)
+	if !closing {
+		// brain is already shutting down
+		return
+	}
+
+	if !b.isHandlingEvents() {
+		// If the event handler loop is not running we must close the inputs
+		// channel from here and drain all pending requests in order to make
+		// b.consumeEvents() exit.
+		close(b.eventsInput)
+		for {
+			select {
+			case _, ok := <-b.eventsLoop:
+				if !ok {
+					// The eventsLoop channel is closed in b.consumeEvents after
+					// all pending messages have been written to it.
+					return
+				}
+			case <-ctx.Done():
+				// shutdown context is expired so we return without waiting for
+				// any pending events.
+				return
+			}
+		}
+	}
+
+	// If we got here then the event handler loop is running and we delegate
+	// proper cleanup and processing of pending messages over there.
+	req := shutdownRequest{
+		ctx:      ctx,
+		callback: make(chan bool),
+	}
+
+	b.shutdown <- req
+	<-req.callback
 }
 
 // Set is a wrapper around the Brains Memory.Set function to allow concurrent
@@ -264,17 +395,6 @@ func (b *Brain) Memories() (map[string]string, error) {
 	b.mu.RUnlock()
 
 	return data, err
-}
-
-// Close is a wrapper around the Brains Memory.Close function to allow
-// concurrent access.
-func (b *Brain) Close() error {
-	b.mu.Lock()
-	b.logger.Debug("Shutting down brain")
-	err := b.memory.Close()
-	b.mu.Unlock()
-
-	return err
 }
 
 func checkHandlerParams(handlerFunc reflect.Type) (evtType reflect.Type, withContext bool, err error) {